+++ qml/DecoLog/Main.qml
@@ -96,6 +96,10 @@ ApplicationWindow {
         else if (what[0] === "setup") { setupDialog.page = parseInt(what[1] || "3"); setupDialog.open() }
         else if (what[0] === "menu") logbook.showMenu(what[1])
         else if (what[0] === "select") logbook.showSelection(what[1], what[2])
+        // Lavori di manutenzione, utili anche da riga di comando.
+        else if (what[0] === "repair") decolog.repairImportedFields()
+        else if (what[0] === "fillall") decolog.completeMissingFromCallbook()
+        else if (what[0] === "maintenance") { decolog.repairImportedFields(); decolog.completeMissingFromCallbook() }
         else if (what[0] === "tab") bottomTabs.currentTab = parseInt(what[1])
         else if (what[0] === "pop") popWindow.active = true
         else if (what[0] === "cluster") openCluster(parseInt(what[1] || "0"))